layers/common/requirements.txt
pyjwt==2.8.0
cryptography==41.0.7
